refactor(key-value): extract render helpers to KeyValueView (T-4-4)
Move the three render-only free functions (render_delete_confirm_modal,
render_kv_context_menu, icon_button_base) from render.rs to view.rs.

KeyValueView grows from 21 LOC to 301 LOC as the named view-layer
boundary. render.rs shrinks from 1064 to 818 LOC.

KeyValueDocument continues to self-render via impl Render in render.rs,
which calls the helpers from view.rs. State coupling across entity
boundaries is not feasible in GPUI without moving all 40+ document fields;
the extraction is scoped to pure rendering utilities that carry no state.

Resolves the T-4-4 deferral flagged by the verify pass (#4778).
